+++ src/main/java/com/alliex/cvs/service/PointService.java
@@ -28,14 +28,14 @@ public class PointService {
     }
 
     @Transactional
-    public Long updatePointPlus(Long userId, int point) {
+    public Long updatePointPlus(Long userId, Long point) {
         Point pointEntity = pointRepository.findByUserId(userId)
                 .orElseThrow(() -> new IllegalArgumentException("Not found point data. user id : " + userId));
 
-        int originPoint = pointEntity.getPoint();
-        int resultPoint = originPoint + point;
+        Long originPoint = pointEntity.getPoint();
+        Long resultPoint = originPoint + point;
         // 충전한도는 정해지는 대로 변경
-        if(resultPoint > 2000000) {
+        if (resultPoint > 2000000) {
             throw new PointLimitExcessException("The point is too much to charge. point : " + point);
         }
 
@@ -46,13 +46,13 @@ public class PointService {
     }
 
     @Transactional
-    public Long updatePointMinus(Long userId, int point) {
+    public Long updatePointMinus(Long userId, Long point) {
         Point pointEntity = pointRepository.findByUserId(userId)
                 .orElseThrow(() -> new IllegalArgumentException("Not found point data. user id : " + userId));
 
-        int originPoint = pointEntity.getPoint();
-        int resultPoint = originPoint - point;
-        if(resultPoint < 0) {
+        Long originPoint = pointEntity.getPoint();
+        Long resultPoint = originPoint - point;
+        if (resultPoint < 0) {
             throw new PointLimitExcessException("The point is not enough to pay. point : " + point);
         }
